requirements.txt
aiohttp==0.17.3
boto3==1.3.1
botocore==1.4.26
Django~=1.8.0
django-extensions==1.5.7
django-storages==1.1.8
djangorestframework==3.2.3
dj-database-url==0.4.1
gunicorn==19.3.0
psycopg2==2.6.1
flake8==2.4.1
pep8==1.5.7
pyflakes==0.8.1
python-dotenv==0.1.3
requests==2.7.0
hashids==1.1.0
pyimgur==0.5.2
two1==3.4.1
pyyaml==3.11
